requirements.txt
# Core
numpy>=1.24
scipy>=1.10
pandas>=2.0
scikit-learn>=1.3
joblib>=1.3

# Deep learning
torch>=2.0
torchvision>=0.15

# Data loading
mat73>=0.60
h5py>=3.9

# Config & utils
pyyaml>=6.0
tqdm>=4.65

# Visualisation
matplotlib>=3.7
seaborn>=0.12

# Jupyter
jupyter>=1.0
ipywidgets>=8.0

# Testing
pytest>=7.4